Checkpoint fixing tests.
